Move pages into modules
