+++ src/modules/ad/AdPage.js
@@ -1,6 +1,6 @@
 import React from "react";
-import useScrollToTop from "../common/hooks/useScrollToTop";
-import Ad from "../modules/ad/components/Ad";
+import useScrollToTop from "../../common/hooks/useScrollToTop";
+import Ad from "./components/Ad";
 
 function AdPage({ match }) {
     useScrollToTop();
