+++ src/modules/favourites/FavouritesPage.js
@@ -1,12 +1,12 @@
 import React from "react";
-import useDocumentTitle from "../common/hooks/useDocumentTitle";
-import useScrollToTop from "../common/hooks/useScrollToTop";
-import BackLink from "../common/components/backlink/BackLink";
-import {CONTEXT_PATH} from "../common/environment";
-import H1WithAutoFocus from "../common/components/h1WithAutoFocus/H1WithAutoFocus";
-import RequiresAuthentication from "../modules/auth/components/RequiresAuthentication";
-import RequiresHasAcceptedTerms from "../modules/user/components/RequiresHasAcceptedTerms";
-import FavouritesList from "../modules/favourites/components/FavouritesList";
+import useDocumentTitle from "../../common/hooks/useDocumentTitle";
+import useScrollToTop from "../../common/hooks/useScrollToTop";
+import BackLink from "../../common/components/backlink/BackLink";
+import {CONTEXT_PATH} from "../../common/environment";
+import H1WithAutoFocus from "../../common/components/h1WithAutoFocus/H1WithAutoFocus";
+import RequiresAuthentication from "../auth/components/RequiresAuthentication";
+import RequiresHasAcceptedTerms from "../user/components/RequiresHasAcceptedTerms";
+import FavouritesList from "./components/FavouritesList";
 
 function FavouritesPage() {
     useDocumentTitle("Favoritter");
