+++ src/modules/saved-searches/SavedSearchesPage.js
@@ -1,13 +1,13 @@
 import React from "react";
-import {CONTEXT_PATH} from "../common/environment";
-import SavedSearchesList from "../modules/saved-searches/components/SavedSearchesList";
-import RequiresAuthentication from "../modules/auth/components/RequiresAuthentication";
-import RequiresHasAcceptedTerms from "../modules/user/components/RequiresHasAcceptedTerms";
-import useDocumentTitle from "../common/hooks/useDocumentTitle";
-import useScrollToTop from "../common/hooks/useScrollToTop";
-import H1WithAutoFocus from "../common/components/h1WithAutoFocus/H1WithAutoFocus";
-import BackLink from "../common/components/backlink/BackLink";
-import "../modules/saved-searches/components/SavedSearches.less";
+import {CONTEXT_PATH} from "../../common/environment";
+import SavedSearchesList from "./components/SavedSearchesList";
+import RequiresAuthentication from "../auth/components/RequiresAuthentication";
+import RequiresHasAcceptedTerms from "../user/components/RequiresHasAcceptedTerms";
+import useDocumentTitle from "../../common/hooks/useDocumentTitle";
+import useScrollToTop from "../../common/hooks/useScrollToTop";
+import H1WithAutoFocus from "../../common/components/h1WithAutoFocus/H1WithAutoFocus";
+import BackLink from "../../common/components/backlink/BackLink";
+import "./components/SavedSearches.less";
 
 /**
  * Page showing saved searches.
