+++ src/modules/superrask-soknad/TrekkSoknadPage.js
@@ -1,7 +1,7 @@
 import React from "react";
-import useScrollToTop from "../common/hooks/useScrollToTop";
-import TrekkSoknad from "../modules/superrask-soknad/components/TrekkSoknad";
-import useDocumentTitle from "../common/hooks/useDocumentTitle";
+import useScrollToTop from "../../common/hooks/useScrollToTop";
+import TrekkSoknad from "./components/TrekkSoknad";
+import useDocumentTitle from "../../common/hooks/useDocumentTitle";
 
 function TrekkSoknadPage({ match }) {
     useDocumentTitle("Trekk søknad");
